% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_tex_converter.R
\name{sGetTexMatrix}
\alias{sGetTexMatrix}
\title{Function returning a matrix as tex output}
\usage{
sGetTexMatrix(pmatAMatrix, pnOutStartLine = 5, pnEndIgnoreLines = 1,
  pnDigits = 2)
}
\arguments{
\item{pmatAMatrix}{Matrix to be represented in tex format}

\item{pnOutStartLine}{line index where output should start, default = 5}

\item{pnEndIgnoreLines}{number of lines to be ignored at the end of the output, default = 1}

\item{pnDigits}{specify the number of digits to be used}
}
\value{
string containing tex representation of matrix
}
\description{
The matrix specified by pmatAMatrix is converted to TeX
using function xtable::xtable. This produces a tabular TeX-object.
The output from xtable::xtable is converted to a TeX-array. Some
lines at the beginning and at the end are ignored. The number
of lines that are ignored can be specified using the parameters
pnOutStartLine and pnEndIgnoreLines.
}

